finished index change
