+++ src/backend/common/types.h
@@ -391,8 +391,7 @@ enum IndexType {
 
   INDEX_TYPE_BTREE = 1,   // btree
   INDEX_TYPE_BWTREE = 2,  // bwtree
-  INDEX_TYPE_HASH = 3,     // hash
-  INDEX_TYPE_BTREE_PRIMARY = 4 // temporarily used. 
+  INDEX_TYPE_HASH = 3     // hash
 };
 
 enum IndexConstraintType {
