+++ src/backend/index/Makefile.am
@@ -11,8 +11,7 @@ index_FILES = \
 			  backend/index/hash_index.cpp \
 			  backend/index/bwtree.cpp \
 			  backend/index/bwtree_index.cpp \
-        backend/index/btree_primary_index.cpp \
-        backend/index/btree_primary_index_explicit.cpp
+        backend/index/btree_primary_index.cpp
 
 index_INCLUDES = \
 				 -I$(srcdir)/backend/common
