+++ src/backend/index/btree_primary_index.cpp
@@ -263,6 +263,151 @@ BTreePrimaryIndex<KeyType, ValueType, KeyComparator, KeyEqualityChecker>::ScanKe
   return result;
 }
 
+
+
+
+
+///////////////////////////////////////////////////////////////////////////////////////////
+
+template <typename KeyType, typename ValueType, class KeyComparator,
+          class KeyEqualityChecker>
+void
+BTreePrimaryIndex<KeyType, ValueType, KeyComparator, KeyEqualityChecker>::Scan(
+    const std::vector<Value> &values, const std::vector<oid_t> &key_column_ids,
+    const std::vector<ExpressionType> &expr_types,
+    const ScanDirectionType &scan_direction, std::vector<std::shared_ptr<ItemPointerHeader>> &result) {
+  KeyType index_key;
+
+  // Check if we have leading (leftmost) column equality
+  // refer : http://www.postgresql.org/docs/8.2/static/indexes-multicolumn.html
+  oid_t leading_column_id = 0;
+  auto key_column_ids_itr = std::find(
+      key_column_ids.begin(), key_column_ids.end(), leading_column_id);
+
+  // SPECIAL CASE : leading column id is one of the key column ids
+  // and is involved in a equality constraint
+  bool special_case = false;
+  if (key_column_ids_itr != key_column_ids.end()) {
+    auto offset = std::distance(key_column_ids.begin(), key_column_ids_itr);
+    if (expr_types[offset] == EXPRESSION_TYPE_COMPARE_EQUAL) {
+      special_case = true;
+    }
+  }
+
+  LOG_TRACE("Special case : %d ", special_case);
+
+  {
+    index_lock.ReadLock();
+
+    auto scan_begin_itr = container.begin();
+    std::unique_ptr<storage::Tuple> start_key;
+    bool all_constraints_are_equal = false;
+
+    // If it is a special case, we can figure out the range to scan in the index
+    if (special_case == true) {
+      start_key.reset(new storage::Tuple(metadata->GetKeySchema(), true));
+
+      // Construct the lower bound key tuple
+      all_constraints_are_equal = ConstructLowerBoundTuple(
+          start_key.get(), values, key_column_ids, expr_types);
+      LOG_TRACE("All constraints are equal : %d ", all_constraints_are_equal);
+      index_key.SetFromKey(start_key.get());
+
+      index_key.SetFromKey(start_key.get());
+
+      // Set scan begin iterator
+      scan_begin_itr = container.equal_range(index_key).first;
+    }
+
+    switch (scan_direction) {
+      case SCAN_DIRECTION_TYPE_FORWARD:
+      case SCAN_DIRECTION_TYPE_BACKWARD: {
+        // Scan the index entries in forward direction
+        for (auto scan_itr = scan_begin_itr; scan_itr != container.end();
+             scan_itr++) {
+          auto scan_current_key = scan_itr->first;
+          auto tuple =
+              scan_current_key.GetTupleForComparison(metadata->GetKeySchema());
+
+          // Compare the current key in the scan with "values" based on
+          // "expression types"
+          // For instance, "5" EXPR_GREATER_THAN "2" is true
+          if (Compare(tuple, key_column_ids, expr_types, values) == true) {
+            std::shared_ptr<ItemPointerHeader> location_header = scan_itr->second;
+            result.push_back(location_header);
+          } else {
+            // We can stop scanning if we know that all constraints are equal
+            if (all_constraints_are_equal == true) {
+              break;
+            }
+          }
+        }
+
+      } break;
+
+      case SCAN_DIRECTION_TYPE_INVALID:
+      default:
+        throw Exception("Invalid scan direction \n");
+        break;
+    }
+
+    index_lock.Unlock();
+  }
+}
+
+template <typename KeyType, typename ValueType, class KeyComparator,
+          class KeyEqualityChecker>
+void BTreePrimaryIndex<KeyType, ValueType, KeyComparator,
+                                    KeyEqualityChecker>::ScanAllKeys(std::vector<std::shared_ptr<ItemPointerHeader>> &result) {
+  {
+    index_lock.ReadLock();
+
+    auto itr = container.begin();
+
+    // scan all entries
+    while (itr != container.end()) {
+      std::shared_ptr<ItemPointerHeader> location = itr->second;
+      result.push_back(location);
+      itr++;
+    }
+
+    index_lock.Unlock();
+  }
+
+}
+
+/**
+ * @brief Return all locations related to this key.
+ */
+template <typename KeyType, typename ValueType, class KeyComparator,
+          class KeyEqualityChecker>
+void
+BTreePrimaryIndex<KeyType, ValueType, KeyComparator, KeyEqualityChecker>::ScanKey(
+    const storage::Tuple *key, std::vector<std::shared_ptr<ItemPointerHeader>> &result) {
+  KeyType index_key;
+  index_key.SetFromKey(key);
+
+  {
+    index_lock.ReadLock();
+
+    // find the <key, location> pair
+    auto entries = container.equal_range(index_key);
+    for (auto entry = entries.first; entry != entries.second; ++entry) {
+      result.push_back(entry->second);
+    }
+
+    index_lock.Unlock();
+  }
+}
+
+
+///////////////////////////////////////////////////////////////////////////////////////////
+
+
+
+
+
+
 template <typename KeyType, typename ValueType, class KeyComparator,
           class KeyEqualityChecker>
 std::string BTreePrimaryIndex<KeyType, ValueType, KeyComparator,
